Changed the underlying data sturcture of DependencyGraph
There are a couple of advantages over the std::set. First of all we
need that the root node of the tree is the first element in the
sequence and we can't assure this with a set that is usally
implemented as a search tree(maybe also balanced).

Also the std::vector should be more compact in memory than a std::set.